Switch shared IP to dedicated IP and vice versa
Ensure the server component for IIS global dns definitions contain at least the
following:

An "A" record with name "[host_name]" and IP Addres "[ip]"

This is regardless if DNS is implemented.

Ensure also the IIS server component configuration contains the Public Shared
IP address, this will be the "external" ip as to be configured in the dns
records

Note: * records as part of the global dns records are not taken care of so they
will need to changed manually. Main reason is that they are not bound to a
website as such and for DNS purposes only.

This change has impacted quite some areas, any testing is welcome